+++ electron.vite.config.ts
@@ -9,6 +9,14 @@ export default defineConfig({
       alias: {
         '@shared': resolve('src/shared')
       }
+    },
+    build: {
+      rollupOptions: {
+        input: {
+          index: resolve('src/main/index.ts'),
+          'chart-render-worker': resolve('src/main/modules/shared/chart-render-worker.ts')
+        }
+      }
     }
   },
   preload: {